Extract calculation of the communication operator.
